requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
--extra-index-url https://www.piwheels.org/simple

click==8.0.3
    # via pip-tools
gpiod==1.5.0
    # via -r requirements.in
mouseinfo==0.1.3
    # via pyautogui
pep517==0.12.0
    # via pip-tools
pillow==8.4.0
    # via -r requirements.in
pip-tools==6.4.0
    # via -r requirements.in
pyautogui==0.9.53
    # via -r requirements.in
pygetwindow==0.0.9
    # via pyautogui
pymsgbox==1.0.9
    # via pyautogui
pyperclip==1.8.2
    # via mouseinfo
pyrect==0.1.4
    # via pygetwindow
pyscreeze==0.1.28
    # via pyautogui
python3-xlib==0.15
    # via
    #   mouseinfo
    #   pyautogui
pytweening==1.0.4
    # via pyautogui
rpi.gpio==0.7.0
    # via -r requirements.in
six==1.16.0
    # via xlib
tomli==1.2.2
    # via pep517
wheel==0.37.0
    # via pip-tools
xlib==0.21
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
